final/foreclosure/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
PyJWT>=2.8.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.43.0
lifelines>=0.27.0
